refactor: migrate to Vue SFC syntax
