+++ eslint.config.js
@@ -1,11 +1,10 @@
 import antfu from '@antfu/eslint-config'
-import VueVine from '@vue-vine/eslint-config'
 
 export default antfu(
   {
+    vue: true,
     rules: {
       /* Override Antfu's default settings */
     },
   },
-  ...VueVine(), // Load VueVine's ESLint config
 )
