+++ package.json
@@ -11,33 +11,32 @@
     "preview": "vite preview",
     "lint": "eslint src --cache",
     "lint:fix": "eslint src --cache --fix",
-    "type-check": "vue-vine-tsc --noEmit --composite false",
+    "type-check": "vue-tsc --noEmit",
     "only-build": "vite build"
   },
   "dependencies": {
-    "@vueuse/core": "^14.1.0",
+    "@vueuse/core": "^14.3.0",
     "gif.js": "^0.2.0",
-    "motion-v": "^1.7.5",
-    "vue": "^3.5.26",
-    "vue-router": "^4.6.4"
+    "motion-v": "^2.3.0",
+    "vue": "^3.5.39",
+    "vue-router": "^5.1.0"
   },
   "devDependencies": {
-    "@antfu/eslint-config": "^6.7.3",
-    "@iconify-json/hugeicons": "^1.2.22",
+    "@antfu/eslint-config": "^9.1.0",
+    "@iconify-json/hugeicons": "^1.2.30",
     "@iconify-json/mdi": "^1.2.3",
     "@iconify-json/ph": "^1.2.2",
     "@iconify-json/solar": "^1.2.5",
     "@iconify-json/svg-spinners": "^1.2.4",
     "@types/gif.js": "^0.2.5",
-    "@types/node": "^25.0.3",
+    "@types/node": "^26.0.1",
     "@vitejs/devtools": "^0.3.3",
-    "@vue-vine/eslint-config": "^1.1.12",
-    "eslint": "^9.39.2",
+    "@vitejs/plugin-vue": "^6.0.7",
+    "eslint": "^10.6.0",
     "npm-run-all": "^4.1.5",
-    "typescript": "^5.9.3",
-    "unocss": "^66.6.7",
-    "vite": "^8.1.0",
-    "vue-vine": "^1.7.27",
-    "vue-vine-tsc": "^1.7.27"
+    "typescript": "^6.0.3",
+    "unocss": "^66.7.4",
+    "vite": "^8.1.2",
+    "vue-tsc": "^3.3.6"
   }
 }
