+++ src/ui/UiCard.vue
@@ -1,0 +1,5 @@
+<template>
+  <div border border-neutral-200 dark:border-neutral-800 rounded-lg shadow-sm p-4>
+    <slot />
+  </div>
+</template>
