+++ src/utils.ts
@@ -94,7 +94,5 @@ export function getAspect(width: number, height: number): [number, number] {
 }
 
 export function sleep(time: number) {
-  return new Promise<void>(resolve =>
-    setTimeout(() => resolve(), time),
-  )
+  return new Promise<void>(resolve => setTimeout(resolve, time))
 }
